+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.2.0
+
+- Support `timer` and `queue_main` with simple blocks as `Glimmer::LibUI.timer` and `Glimmer::LibUI.queue_main`
+- Support examples/timer.rb
+
 ## 0.1.11
 
 - New examples/login.rb